MR-578 - US 3.12 integration with backend

+++ feature_meet_now_screen/src/main/java/com/meetingroom/feature_meet_now/presentation/fast_booking_fragment/RangeSliderState.kt
@@ -1,8 +1,8 @@
 package com.meetingroom.feature_meet_now.presentation.fast_booking_fragment
 
 data class RangeSliderState(
-    val startLimit: Float,
-    val endLimit: Float,
-    var startSelected: Float,
-    var endSelected: Float
+    val startLimit: Int,
+    val endLimit: Int,
+    var startSelected: Int,
+    var endSelected: Int
 )
